ISSUE-02: moved model_objects.py to package pythonsupermarket. adapted dependent files that use it.
