+++ pythonsupermarket/tests/test_supermarket.py
@@ -1,6 +1,6 @@
 import pytest
 
-from model_objects import Product, SpecialOfferType, ProductUnit
+from pythonsupermarket.model_objects import Product, SpecialOfferType, ProductUnit
 from shopping_cart import ShoppingCart
 from teller import Teller
 from tests.fake_catalog import FakeCatalog
